requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
esios==0.25.0
idna==3.10
libsaas==0.4
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
